Change type from int to double

+++ College Marks Counter/MainPage.xaml.cs
@@ -19,8 +19,8 @@ public partial class MainPage : ContentPage
 
 	private void OnCalculateClicked(object sender, EventArgs e)
 	{
-		int marks1 = int.Parse(Marks1Txt.Text);
-        int marks2 = int.Parse(Marks2Txt.Text);
+        double marks1 = double.Parse(Marks1Txt.Text);
+        double marks2 = double.Parse(Marks2Txt.Text);
 
 		double passingMarks = Math.Abs(((marks1 + marks2) / 2.5) - 40);
 		string displayPassingMarks = string.Format("{0:0.00}", passingMarks);
